+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+- feat: store the pitch names and abbreviations and generate a list of them
+  to pick from, based on the entered percents, when "Make sheets" is clicked
+
 ## 0.1.0
 - feat: calculate total percent
 - feat: add intuitive tabindex